+++ ali/common/src/main/java/com/yanny/ali/plugin/glm/IGlobalLootModifierPlugin.java
@@ -8,7 +8,13 @@ import java.util.Optional;
 import java.util.function.BiFunction;
 
 public interface IGlobalLootModifierPlugin extends IPlugin {
-    void registerGlobalLootModifier(IRegistry registry);
+    /**
+     * @deprecated use {@link #registerGlobalLootModifier(IRegistry, ILootTableIdConditionPredicate)} instead
+     */
+    @Deprecated(forRemoval = true, since = "2.2.0")
+    default void registerGlobalLootModifier(IRegistry registry) {}
+
+    void registerGlobalLootModifier(IRegistry registry, ILootTableIdConditionPredicate predicate);
 
     interface IRegistry {
         <T> void registerGlobalLootModifier(Class<T> type, BiFunction<IServerUtils, T, Optional<ILootModifier<?>>> getter);