+++ ali/common/src/main/java/com/yanny/ali/plugin/mods/ICountModifier.java
@@ -1,8 +1,0 @@
-package com.yanny.ali.plugin.mods;
-
-import com.yanny.ali.api.IServerUtils;
-import com.yanny.ali.plugin.server.EnchantedRanges;
-
-public interface ICountModifier {
-    void applyCountModifier(IServerUtils utils, EnchantedRanges count);
-}