+++ alicompat/common/src/main/java/com/yanny/alicompat/accessor/BaseAccessor.java
@@ -1,0 +1,9 @@
+package com.yanny.alicompat.accessor;
+
+public abstract class BaseAccessor<T> {
+    protected final T parent;
+
+    public BaseAccessor(T parent) {
+        this.parent = parent;
+    }
+}