+++ alicompat/common/src/main/java/com/yanny/alicompat/accessor/ICountModifier.java
@@ -1,0 +1,8 @@
+package com.yanny.alicompat.accessor;
+
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.plugin.server.EnchantedRanges;
+
+public interface ICountModifier {
+    void applyCountModifier(IServerUtils utils, EnchantedRanges count);
+}