+++ alicompat/common/src/main/java/com/yanny/alicompat/accessor/IItemListing.java
@@ -1,0 +1,9 @@
+package com.yanny.alicompat.accessor;
+
+import com.yanny.aci.tooltip.TooltipNode;
+import com.yanny.ali.api.IDataNode;
+import com.yanny.ali.api.IServerUtils;
+
+public interface IItemListing {
+    IDataNode getNode(IServerUtils utils, TooltipNode conditions);
+}