+++ alicompat/fabric/src/compat/farmersdelight/java/com/yanny/alicompat/compat/farmersdelight/FDItemListingAccessor.java
@@ -1,0 +1,25 @@
+package com.yanny.alicompat.compat.farmersdelight;
+
+import com.yanny.aci.tooltip.TooltipNode;
+import com.yanny.ali.api.IDataNode;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.alicompat.accessor.BaseAccessor;
+import com.yanny.alicompat.accessor.ClassAccessor;
+import com.yanny.alicompat.accessor.FieldAccessor;
+import com.yanny.alicompat.accessor.IItemListing;
+import net.minecraft.world.entity.npc.villager.VillagerTrades;
+
+@ClassAccessor("vectorwing.farmersdelight.common.event.VillagerEvents$FDItemListing")
+public class FDItemListingAccessor extends BaseAccessor<VillagerTrades.ItemListing> implements IItemListing {
+    @FieldAccessor
+    private VillagerTrades.ItemListing listing;
+
+    public FDItemListingAccessor(VillagerTrades.ItemListing parent) {
+        super(parent);
+    }
+
+    @Override
+    public IDataNode getNode(IServerUtils utils, TooltipNode conditions) {
+        return utils.getItemListing(utils, listing, conditions);
+    }
+}