+++ alicompat/fabric/src/compat/portinglib/java/com/yanny/alicompat/compat/portinglib/LootTableIdConditionAccessor.java
@@ -1,0 +1,28 @@
+package com.yanny.alicompat.compat.portinglib;
+
+import com.yanny.aci.tooltip.TooltipBuilder;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.language.Lang;
+import com.yanny.alicompat.accessor.BaseAccessor;
+import com.yanny.alicompat.accessor.FieldAccessor;
+import com.yanny.alicompat.accessor.IConditionTooltip;
+import io.github.fabricators_of_create.porting_lib.loot.LootTableIdCondition;
+import net.minecraft.resources.Identifier;
+
+public class LootTableIdConditionAccessor extends BaseAccessor<LootTableIdCondition> implements IConditionTooltip {
+    @FieldAccessor
+    private Identifier targetLootTableId;
+
+    public LootTableIdConditionAccessor(LootTableIdCondition parent) {
+        super(parent);
+    }
+
+    public Identifier getTargetLootTableId() {
+        return targetLootTableId;
+    }
+
+    @Override
+    public TooltipBuilder getTooltip(IServerUtils utils) {
+        return utils.getValueTooltip(utils, targetLootTableId).key(Lang.Conditions.LOOT_TABLE_ID);
+    }
+}